data_management: shorten include files, refactor
