+++ cpp/onedal/detail/helpers.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "onedal/ranges.hpp"
+#include "onedal/types_data.hpp"
 
 namespace dal {
 namespace detail {
@@ -9,6 +9,6 @@ struct table_range {
     range rows;
     range columns;
 };
-    
+
 } // namespace detail
 } // namespace dal
