+++ cpp/onedal/detail/type_rt.hpp
@@ -15,7 +15,7 @@
  *******************************************************************************/
 
 #pragma once
-#include "onedal/integral_types.hpp"
+#include "onedal/types_integral.hpp"
 
 namespace dal {
 namespace detail {
